add: routes
move User endpoints in 'index.js' to 'routes' folder
